% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradford.R
\name{bradford}
\alias{bradford}
\alias{dbradford}
\alias{pbradford}
\alias{qbradford}
\alias{rbradford}
\title{The Standardized Bradford Distribution}
\source{
The main definition for for the Standardized Bradford Function was taken from
\href{https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.bradford.html}{SciPy}.
The \code{rbradford} function is defined using the \href{https://en.wikipedia.org/wiki/Inverse_transform_sampling}{Inverse Transform Sampling Method,}
with a more in depth explanation presented in the package's vignettes. All functions
were built to work like the probability functions in the\href{https://www.rdocumentation.org/packages/stats/versions/3.6.2}{\code{{stats}}} package.
}
\usage{
dbradford(x, c = 5, log = FALSE)

pbradford(q, c = 5, lower.tail = TRUE, log.p = FALSE)

qbradford(p, c = 5, lower.tail = TRUE, log.p = FALSE)

rbradford(n, c = 5)
}
\arguments{
\item{x, q}{A vector of quantiles. Must be between 0 and 1.}

\item{c}{The parameter of the standardized bradford distribution. Must be
greater than 0.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p)}

\item{lower.tail}{logical; if TRUE (default), probabilities are given as
P(x<=X); otherwise, P(x > X).}

\item{p}{A vector of probabilities.}

\item{n}{number of observations. If \code{length(n)>1}, the length is taken to be
the number required.}
}
\value{
A numeric vector.
}
\description{
Density, distribution function, quantile function, and random generation for
the Standardized Bradford Distribuction with parameter equal to c, which must
be greater than 0. \code{dbradford} gives the density, \code{pbradford} gives the distribution function,
\code{qbradford} gives the quantile function and \code{rnorm} generates random numbers
from the distribution.
}
\details{
If c is not specified, it assumes the default value of 5.
The Standardized Bradford Distribution has density defined as
\deqn{f(x;c) = \frac{c}{log(1+c) log(1+ cx)}}
for \eqn{c > 0} and \eqn{0 \leq x \leq 1}.
}
\examples{
set.seed(1996)

# Generate 10 random numbers with c=10
rbradford(n=10, c=10)

# What value satisfies P(X <= x) = 0.5 when c=13?
qbradford(p=0.5, c=13)

# Probability of P(X <=0.5) when c=10
pbradford(q=0.5, c=10)

# Density function of x=0.375 when c=100
dbradford(x=0.375, c=100)
}
