% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbradford.R
\name{dbradford}
\alias{dbradford}
\title{dbradford}
\usage{
dbradford(x, c = 5, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles. Must be between 0 and 1.}

\item{c}{parameter of the standardized bradford distribution. Must be
greater than 0.}

\item{log}{logical; if TRUE, probabilities p are given as log(p)}
}
\value{
a vector of \code{length(x)} with distribution probabilities.
}
\description{
This function returns the distribution of the Standardized Bradford Distribution.
}
\examples{
dbradford(x=0.375, c=100)

}
