% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbradford.R
\name{pbradford}
\alias{pbradford}
\title{pbradford}
\usage{
pbradford(q, c = 5, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{vector of quantiles. Must be between 0 and 1.}

\item{c}{parameter of the standardized bradford distribution. Must be
greater than 0.}

\item{lower.tail}{logical; if TRUE (default), probabilities are given as
P(x<=X); otherwise, P(x > X).}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p)}
}
\value{
a vector of \code{length(q)} with accumulated distribution probabilities.
}
\description{
This function returns probabilities from the accumulated distribution function
of the Bradford Distribution.
}
\examples{
pbradford(q=0.5, c=10)

}
