% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbradford.R
\name{qbradford}
\alias{qbradford}
\title{qbradford}
\usage{
qbradford(p, c = 5, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{c}{parameter of the standardized bradford distribution. Must be
greater than 0.}

\item{lower.tail}{logical; if TRUE (default), probabilities are given as
P(x<=X); otherwise, P(x > X).}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p)}
}
\value{
a vector of \code{length(p)} with the distribution quantiles.
}
\description{
This function gives the quantile function from the Standardized
Bradford Distribution.
}
\examples{
qbradford(p=0.5, c=13)
}
