% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbradford.R
\name{rbradford}
\alias{rbradford}
\title{rbradford}
\usage{
rbradford(n, c = 5)
}
\arguments{
\item{n}{number of observations. If \code{length(n)>1}, the length is taken to be
the number required.}

\item{c}{parameter of the standardized bradford distribution. Must be
greater than 0.}
}
\value{
a vector of length \code{n} with values generated from the Standardized
Bradford Distribution with parameter \code{c}.
}
\description{
This function generates a vector of values from the Standardized Bradford
Distribution with length \code{n} and parameter \code{c}.
}
\examples{
rbradford(n=100, c=10)

}
